added responsive graph size
This change was done by using a callback event (on resize) in d3. Initial
dimentions was a bit tricky so it was removed -- instead, the aforementioned
event is triggered manually in the parent component. As the dimentions are
now dynamic, the width & height stuff in graph/config.js had to be removed,
and cascading changes had to be done in graph/d3state.js.

Altohugh everything seems to work, the implementation isn't perfect:
as the window gets smaller, the bounds get more inaccurate, presumably
because of style margins? Also, the node group and force config is
recalculated on each resize + the apply func in d3state.js is called,
which might be a bit inefficient. Also, the code is a bit messy.