fixed pan jitter
Apparently, the jitters are caused because of how the panning
action does a mouse position read while applying the changes,
basically. Not sure why this is a problem but it got fixed by
wrapping the main svg with another layer (dedicated for ui).
zoom is still scruffy (but not becase of jitters).